+++ CHANGELOG.md
@@ -2,5 +2,12 @@
 
 All notable changes to `laravel-pest-recorder` will be documented in this file.
 
+## 1.0.1 - 2026-03-05
+* Ensure Browser tests directory exists
+* Implement assertText -> assertSee method
+* Implement --visit=... option
+* Fix for clicking buttons without any id or test-id
+* Implement --acting-as=user option
+
 ## 1.0.0 - 2026-02-14
 * Initial release